+++ CHANGELOG.md
@@ -11,6 +11,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Instance based and/or machine independent RustSBI to support hypervisor development
 - Structure `MachineInfo` for non-machine environment, e.g. cross-architecture emulator
+- Builder structure for instance based RustSBI framework
+- Implement RustSBI traits for their references
 - Feature `legacy` to gate SBI legacy extension
 - Expose `init_*` functions on instance based RustSBI implementation
 - LEGACY_CLEAR_IPI implemented