+++ CHANGELOG.md
@@ -9,6 +9,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+### Modified
+
+### Removed
+
+## [0.3.0] - 2022-11-03
+
+### Added
+
 - Instance based and/or machine independent RustSBI to support hypervisor development
 - Structure `MachineInfo` for non-machine environment, e.g. cross-architecture emulator
 - Builder structure for instance based RustSBI framework
@@ -148,7 +156,8 @@ RustSBI is adapted to SBI standard with implementation number 4.
 - RISC-V Privileged Specification v1.11
 - Backward compatible to privileged spec v1.9.1
 
-[Unreleased]: https://github.com/rustsbi/rustsbi/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/rustsbi/rustsbi/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/rustsbi/rustsbi/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/rustsbi/rustsbi/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/rustsbi/rustsbi/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/rustsbi/rustsbi/compare/v0.1.1...v0.2.0